1.use chan write access log msg to file
