+++ cmd/proxy/proxy.go
@@ -18,6 +18,7 @@
 package main
 
 import (
+	_ "net/http/pprof"
 	"os"
 	"strconv"
 	"time"
@@ -26,10 +27,10 @@ import (
 import (
 	_ "github.com/apache/dubbo-go/metadata/service/inmemory"
 	"github.com/urfave/cli"
-	_ "net/http/pprof"
 )
 
 import (
+	_ "github.com/dubbogo/dubbo-go-proxy/pkg/filter/accesslog"
 	_ "github.com/dubbogo/dubbo-go-proxy/pkg/filter/logger"
 	_ "github.com/dubbogo/dubbo-go-proxy/pkg/filter/recovery"
 	_ "github.com/dubbogo/dubbo-go-proxy/pkg/filter/remote"
