+++ pkg/model/http.go
@@ -24,7 +24,6 @@ type HttpConnectionManager struct {
 	HTTPFilters       []HTTPFilter           `yaml:"http_filters" json:"http_filters" mapstructure:"http_filters"`
 	ServerName        string                 `yaml:"server_name" json:"server_name" mapstructure:"server_name"`
 	IdleTimeoutStr    string                 `yaml:"idle_timeout" json:"idle_timeout" mapstructure:"idle_timeout"`
-	AccessLog         AccessLog              `yaml:"access_log" json:"access_log" mapstructure:"access_log"`
 	GenerateRequestID bool                   `yaml:"generate_request_id" json:"generate_request_id" mapstructure:"generate_request_id"`
 }
