+++ pkg/proxy/listener.go
@@ -152,7 +152,10 @@ func (s *DefaultHttpListener) ServeHTTP(w http.ResponseWriter, r *http.Request) 
 func addFilter(ctx *h.HttpContext, api router.API) {
 	ctx.AppendFilterFunc(extension.GetMustFilterFunc(constant.LoggerFilter),
 		extension.GetMustFilterFunc(constant.RecoveryFilter), extension.GetMustFilterFunc(constant.TimeoutFilter))
-
+	alc := config.GetBootstrap().StaticResources.AccessLogConfig
+	if alc.Enable {
+		ctx.AppendFilterFunc(extension.GetMustFilterFunc(constant.AccessLogFilter))
+	}
 	switch api.Method.IntegrationRequest.RequestType {
 	// TODO add some basic filter for diff protocol
 	case config.DubboRequest:
